feat: add laravel pint

+++ app/Domains/Budget/Imports/BudgetImport.php
@@ -7,20 +7,18 @@ use App\Domains\Imports\ImportConcern;
 use Carbon\Carbon;
 use Insane\Journal\Models\Core\Category;
 
-
 class BudgetImport extends ImportConcern
 {
     /**
-    * @param array $row
-    *
-    * @return \Illuminate\Database\Eloquent\Model|null
-    */
+     * @return \Illuminate\Database\Eloquent\Model|null
+     */
     public function model(array $row)
     {
         return BudgetMonth::createBudget($row);
     }
 
-    public function getYNABMoney($currencyValue)  {
+    public function getYNABMoney($currencyValue)
+    {
         $CODES = [
             'RD' => 'DOP',
             'USD' => 'USD',
@@ -28,15 +26,14 @@ class BudgetImport extends ImportConcern
         $inflow = explode('$', str_replace('-', '', $currencyValue));
         $money = floatval(preg_replace('/[^-\d\.]/', '', $currencyValue));
         $amounts = (object) [
-            'money' => (double) $money,
+            'money' => (float) $money,
             'currency_code' => $CODES[$inflow[0]] ?? 'USD',
         ];
 
-
-       return [
+        return [
             'amount' => $amounts->money,
             'currency_code' => $amounts->currency_code,
-       ];
+        ];
     }
 
     public function map($row): array
@@ -47,7 +44,7 @@ class BudgetImport extends ImportConcern
 
         $categoryGroupId = Category::findOrCreateByName($this->session, $row['category_group']);
         $transactionCategoryId = Category::findOrCreateByName($this->session, $row['category'], $categoryGroupId);
-        $month = Carbon::createFromFormat('M Y', $row['month'])->format('Y-m') . "-01";
+        $month = Carbon::createFromFormat('M Y', $row['month'])->format('Y-m').'-01';
 
         $row['team_id'] = $this->session['team_id'];
         $row['user_id'] = $this->session['user_id'];
@@ -59,10 +56,11 @@ class BudgetImport extends ImportConcern
         $row['activity'] = $activity['amount'];
         $row['available'] = $available['amount'];
         $row['metaData'] = [
-            "resource_id" => 'YNAB',
-            "resource_origin" => 'YNAB',
-            "resource_type" => 'budget',
+            'resource_id' => 'YNAB',
+            'resource_origin' => 'YNAB',
+            'resource_type' => 'budget',
         ];
+
         return $row;
     }
 }
